+++ controllers/LoginController.php
@@ -1,7 +1,11 @@
 <?php
+
 namespace app\controllers;
+
 use app\models\LoginForm;
 use yii\web\Controller;
+use yii\web\Response;
+use yii\widgets\ActiveForm;
 
 class LoginController extends Controller
 {
@@ -11,16 +15,21 @@ class LoginController extends Controller
     {
         $loginForm = new LoginForm();
 
-        if (\Yii::$app->request->getIsPost()) {
-            $loginForm->load(\Yii::$app->request->post());
+        if (\Yii::$app->request->getIsPost() && $loginForm->load(\Yii::$app->request->post())) {
+            if (\Yii::$app->request->isAjax && $loginForm->load(\Yii::$app->request->post())) {
+                \Yii::$app->response->format = Response::FORMAT_JSON;
 
+                return ActiveForm::validate($loginForm);
+            }
             if ($loginForm->validate()) {
+
                 $user = $loginForm->getUser();
                 \Yii::$app->user->login($user);
 
                 return $this->goHome();
             }
         }
+
 
         return $this->render('index', ['loginForm' => $loginForm]);
     }
