--- models/UserLogin.php
@@ -1,54 +1,0 @@
-<?php
-namespace app\models;
-use yii\db\ActiveRecord;
-use yii\web\IdentityInterface;
-
-class UserLogin extends ActiveRecord implements IdentityInterface
-{
-
-    public static function findIdentity($id)
-    {
-        return self::findOne($id);
-    }
-
-    public static function findIdentityByAccessToken($token, $type = null)
-    {
-        // TODO: Implement findIdentityByAccessToken() method.
-    }
-
-    public function getId()
-    {
-        return $this->getPrimaryKey();
-    }
-
-    public function getAuthKey()
-    {
-        // TODO: Implement getAuthKey() method.
-    }
-
-    public function validateAuthKey($authKey)
-    {
-        // TODO: Implement validateAuthKey() method.
-    }
-
-    public function attributeLabels()
-    {
-        return [
-            'email' => 'Электронная почта',
-            'password' => 'Пароль',
-        ];
-    }
-
-    public function validatePassword($password)
-    {
-        return \Yii::$app->security->validatePassword($password, $this->password);
-    }
-
-    public function rules()
-    {
-        return [
-            [['email', 'password'], 'safe'],
-        ];
-    }
-
-}
