+++ views/layouts/landing.php
@@ -212,14 +212,14 @@ app\assets\LandingAsset::register($this);
                 </div>
             </div>
         </footer>
-        <section class="modal enter-form form-modal" id="enter-form">
-            <h2>Вход на сайт</h2>
+      <!--  <section class="modal enter-form form-modal" id="enter-form">
+            <h2>Вход на сайт</h2>-->
 
 
             <?php echo $content; ?>
 
 
-            <button class="form-modal-close" type="button">Закрыть</button>
+        <!--    <button class="form-modal-close" type="button">Закрыть</button>-->
         </section>
     </div>
     <div class="overlay"></div>
